Stable working version

+++ TrainCat/SpriteUtils.m
@@ -15,7 +15,7 @@
 {
     CCSprite *sprite = [CCSprite node];
     GLubyte *buffer = malloc(sizeof(GLubyte)*4);
-    for (int i=0;i<4;i++) {buffer[i]=255;}
+    for (NSUInteger i=0;i<4;i++) {buffer[i]=255;}
     CCTexture2D *tex = [[CCTexture2D alloc] initWithData:buffer pixelFormat:kCCTexture2DPixelFormat_RGB5A1 pixelsWide:1 pixelsHigh:1 contentSize:size];
     [sprite setTexture:tex];
     [sprite setTextureRect:CGRectMake(0, 0, size.width, size.height)];
